Move software inventory helpers into software management
